Use symlinks to denote to-be-linked files in initializeUpload.
This involves the addition of a new utility to create placeholder symlinks. If
these symlinks are detected in the list of files given to initializeUpload, the
function will automatically generate deduplicating links rather than uploading
them as regular files. This mechanism avoids the need to store external linking
info to be passed to dedup.link= in initializeUpload.