Prevent any linking from metadata documents.
The metadata itself doesn't have a MD5 sum, so we can't use it in MD5-based
deduplication anyway. By also blocking explicit links, we simplify the
situation as we can always assume that the metadata is a real JSON document
rather than a link or some other weird thing - for example, how do we check
the MD5 sum of an artifact if its metadata document is a placeholder?